Refactor: Move Controllers for Web to App\Http\Controllers\Web"
