+++ app/Http/Controllers/Web/OrderController.php
@@ -1,10 +1,11 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Web;
 
+use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 
-class OrderItemController extends Controller
+class OrderController extends Controller
 {
     /**
      * Display a listing of the resource.
