+++ app/Http/Controllers/Web/OrderItemController.php
@@ -1,10 +1,11 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Web;
 
+use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 
-class UserController extends Controller
+class OrderItemController extends Controller
 {
     /**
      * Display a listing of the resource.
